+++ pkg/p2p/libp2p/internal/reacher/reacher_test.go
@@ -12,7 +12,7 @@ import (
 	"time"
 
 	"github.com/ethersphere/bee/pkg/p2p"
-	"github.com/ethersphere/bee/pkg/p2p/libp2p/reacher"
+	"github.com/ethersphere/bee/pkg/p2p/libp2p/internal/reacher"
 	"github.com/ethersphere/bee/pkg/swarm"
 	"github.com/ethersphere/bee/pkg/swarm/test"
 	ma "github.com/multiformats/go-multiaddr"
